backend/requirements/requirements.txt
appdirs==1.4.3
Django==1.9.9
django-environ==0.4.3
djangorestframework==3.6.2
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
six==1.10.0
